prometheus-grafana/otel/python-app-otlp/app/requirements.txt
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-sdk
